requirements.txt
transformers
datasets
accelerate==0.27.0
deepspeed==0.14.0
evaluate
matplotlib
hydra-core
omegaconf
peft
rouge_score
tqdm
matplotlib
einops
packaging
bitsandbytes
scipy
ninja
sentencepiece
protobuf
wandb
google-generativeai
openai
aiohttp_jinja2
huggingface_hub
scikit-learn
numpy==1.23.5
flash-attn==2.5.8